fix(permissions): Fix permission requesting
The problem that the user needed to press twice the "Request Permission"
button for the Bluetooth connection is fixed by using
rememberMultiplePermissionsState.
Now the app instantly asks for every permission it needs and if the user
declines one, he is able to press the button to request the declined
ones.